split Fs stories and fix scrolling
